pyproject.toml
[project]
name = "photo2stitch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "numpy",
    "opencv-python",
    "pillow",
    "scikit-learn",
]

[project.scripts]
photo2stitch = "photo2stitch.__main__:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"
